% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-var-group.R
\name{get_member}
\alias{get_member}
\title{Access a single member of a \code{survey_var_group}}
\usage{
get_member(g, option)
}
\arguments{
\item{g}{A \link{survey_var_group} object.}

\item{option}{A single string, one of \code{get_group_options(g)}.}
}
\value{
The \link{survey_var} for that option.
}
\description{
Access a single member of a \code{survey_var_group}
}
