% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-var.R
\name{survey_var-accessors}
\alias{survey_var-accessors}
\alias{get_stem}
\alias{get_cols}
\alias{get_type}
\alias{get_question_text}
\alias{get_subquestions}
\alias{get_option_labels}
\alias{get_response_options}
\title{Access \code{survey_var} fields}
\usage{
get_stem(v)

get_cols(v)

get_type(v)

get_question_text(v)

get_subquestions(v)

get_option_labels(v)

get_response_options(v)
}
\arguments{
\item{v}{A \link{survey_var} object.}
}
\value{
The requested field.
}
\description{
Access \code{survey_var} fields
}
