% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-var-group.R
\name{survey_var_group-accessors}
\alias{survey_var_group-accessors}
\alias{get_group_var}
\alias{get_group_options}
\alias{get_members}
\title{Access \code{survey_var_group} fields}
\usage{
get_group_var(g)

get_group_options(g)

get_members(g)
}
\arguments{
\item{g}{A \link{survey_var_group} object.}
}
\value{
The requested field.
}
\description{
Access \code{survey_var_group} fields
}
